Affected Issue(s): Move QR code creation to "ANC visit reminder" flow
https://www.notion.so/Back-End-Move-the-QR-code-generation-to-ANC-Visit-Reminder-flow-96f94c9065704ac9ae36f29018296d70

What this commit has achieved:
1. Refactored ANC Visit Reminder service, added functionality to add
parameters with header (image) for QR code.
2. Renamed test `QRCodeGapCareServiceTest` to `FhirResourceServiceTest`
3. Modified queries for ANC Visit Reminder, added a new column retrieved
`pregnancyGapCommaSeparatedValues`
