+++ src/main/java/org/sidindonesia/bidanreport/integration/qontak/config/property/QontakProperties.java
@@ -29,6 +29,7 @@ public class QontakProperties {
 		private String pregnantWomanMessageTemplateId;
 		private String nonPregnantWomanMessageTemplateId;
 		private String visitReminderMessageTemplateId;
+		private String visitReminderWithHeaderImageMessageTemplateId;
 		private Integer visitReminderIntervalInDays;
 		private Integer visitIntervalInDays;
 		private String pregnancyGapMessageTemplateId;
