+++ src/main/java/org/sidindonesia/bidanreport/repository/constant/QueryConstants.java
@@ -95,7 +95,17 @@ public final class QueryConstants {
 		+ "    AND av_sub2.anc_date = av_max_anc_date.latest_anc_date) " + "  GROUP BY "
 		+ "   av_sub2.mother_base_entity_id, " + "   av_sub2.anc_date) av_max_event_id ON "
 		+ "  av_sub1.event_id = av_max_event_id.latest_event_id " + " WHERE "
-		+ "  av_sub1.mother_base_entity_id = mi.mother_base_entity_id) AS latestAncVisitNumber " + "FROM "
+		+ "  av_sub1.mother_base_entity_id = mi.mother_base_entity_id) AS latestAncVisitNumber, " + " ( " + SELECT + PREGNANCY_GAP_COMMA_SEPARATED_VALUES + FROM
+		+ "  {h-schema}anc_visit av_sub1 " + " INNER JOIN ( " + SELECT + "   MAX(av_sub2.event_id) AS latest_event_id "
+		+ FROM + "   {h-schema}anc_visit av_sub2 " + "  INNER JOIN ( " + "   SELECT " + "    mother_base_entity_id, "
+		+ "    MAX(anc_date) AS latest_anc_date " + "   FROM " + "    {h-schema}anc_visit av_sub3 " + "   GROUP BY "
+		+ "    mother_base_entity_id) av_max_anc_date ON "
+		+ "   (av_sub2.mother_base_entity_id = av_max_anc_date.mother_base_entity_id "
+		+ "    AND av_sub2.anc_date = av_max_anc_date.latest_anc_date) " + "  GROUP BY "
+		+ "   av_sub2.mother_base_entity_id, " + "   av_sub2.anc_date) av_max_event_id ON "
+		+ "  av_sub1.event_id = av_max_event_id.latest_event_id " + WHERE
+		+ "  av_sub1.mother_base_entity_id = mi.mother_base_entity_id "
+		+ "  AND av_sub1.date_created >= current_date - INTERVAL '1 day' * ?1) AS pregnancyGapCommaSeparatedValues " + "FROM "
 		+ " {h-schema}mother_identity mi " + "WHERE " + " mi.event_id IN ( " + " SELECT "
 		+ "  MAX(mi_id_only.event_id) OVER (PARTITION BY mi_id_only.mobile_phone_number) " + " FROM "
 		+ "  {h-schema}mother_identity mi_id_only " + " INNER JOIN {h-schema}anc_register ar ON "
@@ -131,7 +141,17 @@ public final class QueryConstants {
 		+ "    AND av_sub2.anc_date = av_max_anc_date.latest_anc_date) " + "  GROUP BY "
 		+ "   av_sub2.mother_base_entity_id, " + "   av_sub2.anc_date) av_max_event_id ON "
 		+ "  av_sub1.event_id = av_max_event_id.latest_event_id " + " WHERE "
-		+ "  av_sub1.mother_base_entity_id = me.mother_base_entity_id) AS latestAncVisitNumber " + "FROM "
+		+ "  av_sub1.mother_base_entity_id = me.mother_base_entity_id) AS latestAncVisitNumber, " + " ( " + SELECT + PREGNANCY_GAP_COMMA_SEPARATED_VALUES + FROM
+		+ "  {h-schema}anc_visit av_sub1 " + " INNER JOIN ( " + SELECT + "   MAX(av_sub2.event_id) AS latest_event_id "
+		+ FROM + "   {h-schema}anc_visit av_sub2 " + "  INNER JOIN ( " + "   SELECT " + "    mother_base_entity_id, "
+		+ "    MAX(anc_date) AS latest_anc_date " + "   FROM " + "    {h-schema}anc_visit av_sub3 " + "   GROUP BY "
+		+ "    mother_base_entity_id) av_max_anc_date ON "
+		+ "   (av_sub2.mother_base_entity_id = av_max_anc_date.mother_base_entity_id "
+		+ "    AND av_sub2.anc_date = av_max_anc_date.latest_anc_date) " + "  GROUP BY "
+		+ "   av_sub2.mother_base_entity_id, " + "   av_sub2.anc_date) av_max_event_id ON "
+		+ "  av_sub1.event_id = av_max_event_id.latest_event_id " + WHERE
+		+ "  av_sub1.mother_base_entity_id = me.mother_base_entity_id "
+		+ "  AND av_sub1.date_created >= current_date - INTERVAL '1 day' * ?1) AS pregnancyGapCommaSeparatedValues " + "FROM "
 		+ " {h-schema}mother_edit me " + "WHERE " + " me.event_id IN ( " + " SELECT "
 		+ "  MAX(me_id_only.event_id) OVER (PARTITION BY me_id_only.mother_base_entity_id) " + " FROM "
 		+ "  {h-schema}mother_edit me_id_only " + " INNER JOIN {h-schema}anc_register ar ON "
