+++ src/main/java/org/sidindonesia/bidanreport/repository/projection/AncVisitReminderProjection.java
@@ -2,4 +2,5 @@ package org.sidindonesia.bidanreport.repository.projection;
 
 public interface AncVisitReminderProjection extends MotherIdentityWhatsAppProjection {
 	Integer getLatestAncVisitNumber();
+	String getPregnancyGapCommaSeparatedValues();
 }
